GASNet fork to be built on Darwin cluster
The most important part is in Makefile.am.mxm_conduit.darwin file - it is patched to use configure's MPICC/MPICXX instead of environment MPICC/MPICXX. This is done because we should tell configure to use -fPIC for relocatable code and, by default, environment MPI compiler does not use that variable.

This is needed for linking Regent's liblegion_terra.so.

Makefile.darwin is a simplified main Makefile to be used only on Darwin.

configs/config.darwin.release is a configure's options set for Darwin.